Properly nest returning judgements for a book under the /book/:book_id/judgements route
