+++ app/controllers/api/v1/books_controller.rb
@@ -4,7 +4,6 @@ require 'csv'
 
 module Api
   module V1
-    # rubocop:disable Metrics/ClassLength
     class BooksController < Api::ApiController
       before_action :set_book, only: [ :show, :update, :destroy ]
       before_action :check_book, only: [ :show, :update, :destroy ]
@@ -27,63 +26,13 @@ module Api
         respond_with @books
       end
 
-      # rubocop:disable Metrics/MethodLength
-      # rubocop:disable Metrics/AbcSize
-      # rubocop:disable Metrics/CyclomaticComplexity
-      # rubocop:disable Metrics/PerceivedComplexity
-      # rubocop:disable Metrics/BlockLength
       api :GET, '/api/books/:book_id',
           'Show the book with the given ID.'
       param :id, :number,
             desc: 'The ID of the requested book.', required: true
       def show
-        respond_to do |format|
-          format.json
-          format.csv do
-            @csv_array = []
-            csv_headers = %w[query docid]
-
-            # Only return rateable judgements, filter out the unrateable ones.
-            # unique_raters = @book.judgements.rateable.preload(:user).collect(&:user).uniq
-            # unique_raters = @book.judges.merge(Judgement.rateable)
-            unique_judge_ids = @book.query_doc_pairs.joins(:judgements)
-              .distinct.pluck(:user_id)
-
-            # this logic about using email versus name is kind of awful.  Think about user.full_name or user.identifier?
-            unique_judges = []
-            unique_judge_ids.each do |judge_id|
-              judge = User.find(judge_id) unless judge_id.nil?
-              unique_judges << judge
-              csv_headers << make_csv_safe(if judge.nil?
-                                             'anonymous'
-                                           else
-                                             judge.name.presence || judge.email
-                                           end)
-            end
-
-            @csv_array << csv_headers
-            query_doc_pairs = @book.query_doc_pairs.includes(:judgements)
-            query_doc_pairs.each do |qdp|
-              row = [ make_csv_safe(qdp.query_text), qdp.doc_id ]
-              unique_judges.each do |judge|
-                judgement = qdp.judgements.detect { |j| j.user == judge }
-                rating = judgement.nil? ? '' : judgement.rating
-
-                row.append rating
-              end
-              @csv_array << row
-            end
-
-            headers['Content-Disposition'] = "attachment; filename=\"book_#{@book.id}_export.csv\""
-            headers['Content-Type'] ||= 'text/csv'
-          end
-        end
+        respond_with @judgement
       end
-      # rubocop:enable Metrics/MethodLength
-      # rubocop:enable Metrics/AbcSize
-      # rubocop:enable Metrics/CyclomaticComplexity
-      # rubocop:enable Metrics/PerceivedComplexity
-      # rubocop:enable Metrics/BlockLength
 
       api :POST, '/api/books', 'Create a new book.'
       param_group :book_params
@@ -139,15 +88,6 @@ module Api
       def check_book
         render json: { message: 'Book not found!' }, status: :not_found unless @book
       end
-
-      def make_csv_safe str
-        if %w[- = + @].include?(str[0])
-          " #{str}"
-        else
-          str
-        end
-      end
     end
-    # rubocop:enable Metrics/ClassLength
   end
 end
