+++ app/controllers/api/v1/judgements_controller.rb
@@ -2,6 +2,7 @@
 
 module Api
   module V1
+    # rubocop:disable Metrics/ClassLength
     class JudgementsController < Api::ApiController
       before_action :set_book
       before_action :check_book
@@ -29,11 +30,76 @@ module Api
         end
       end
 
+      # rubocop:disable Metrics/MethodLength
+      # rubocop:disable Metrics/AbcSize
+      # rubocop:disable Metrics/CyclomaticComplexity
+      # rubocop:disable Metrics/PerceivedComplexity
+      # rubocop:disable Metrics/BlockLength
       api :GET, '/api/books/:book_id/judgements',
-          'List all judgements for the book'
+          'List all judgements for the book.  When you request the .csv version we only return valid rated judgements.'
       param :book_id, :number,
             desc: 'The ID of the requested book.', required: true
       def index
+        respond_to do |format|
+          format.json do
+            @judgements = @book.judgements
+          end
+          format.csv do
+            @csv_array = []
+            csv_headers = %w[query docid]
+
+            # can't use more efficient quering because we need to include judgements where the user_id is nil.
+            # Maybe someday we require a user for all judgements?  Even if the user has the name 'Anonymous'?
+            # Only return rateable judgements, filter out the unrateable ones.
+            # unique_judge_ids = @book.query_doc_pairs.joins(:judgements).rateable
+            #  .distinct.pluck(:user_id)
+            # unique_judge_ids = @book.judgements.rateable.select(:user_id).distinct.pluck(:user_id)
+            unique_judge_ids = @book.query_doc_pairs.joins(:judgements)
+              .distinct.pluck(:user_id)
+
+            # unique_judges = User.where(id: @book.judgements.select(:user_id).distinct)
+            # unique_judges.each do |judge|
+            #  csv_headers << make_csv_safe(judge.fullname)
+            # end
+            unique_judges = []
+            unique_judge_ids.each do |judge_id|
+              judge = User.find(judge_id) unless judge_id.nil?
+              unique_judges << judge
+              csv_headers << make_csv_safe(if judge.nil?
+                                             'Anonymous'
+                                           else
+                                             judge.fullname
+                                           end)
+            end
+
+            @csv_array << csv_headers
+            query_doc_pairs = @book.query_doc_pairs
+              .joins(:judgements)
+              .where.not(judgements: { rating: nil })
+              .includes(:judgements)
+            query_doc_pairs.each do |qdp|
+              row = [ make_csv_safe(qdp.query_text), qdp.doc_id ]
+              unique_judges.each do |judge|
+                judgement = qdp.judgements.detect { |j| j.user == judge }
+                rating = judgement.nil? ? '' : judgement.rating
+
+                row.append rating
+              end
+              @csv_array << row
+            end
+
+            headers['Content-Disposition'] = "attachment; filename=\"book_#{@book.id}_judgements.csv\""
+            headers['Content-Type'] ||= 'text/csv'
+          end
+        end
+      end
+      # rubocop:enable Metrics/MethodLength
+      # rubocop:enable Metrics/AbcSize
+      # rubocop:enable Metrics/CyclomaticComplexity
+      # rubocop:enable Metrics/PerceivedComplexity
+      # rubocop:enable Metrics/BlockLength
+
+      def indexold
         @judgements = @book.judgements
 
         respond_with @judgements
@@ -108,6 +174,14 @@ module Api
 
       private
 
+      def make_csv_safe str
+        if %w[- = + @].include?(str[0])
+          " #{str}"
+        else
+          str
+        end
+      end
+
       def judgement_params
         params.require(:judgement).permit(:rating, :unrateable, :query_doc_pair_id, :user_id, :explanation)
       end
@@ -120,5 +194,6 @@ module Api
         render json: { message: 'Query Doc Pair not found!' }, status: :not_found unless @judgement
       end
     end
+    # rubocop:enable Metrics/ClassLength
   end
 end
