+++ test/controllers/api/v1/books_controller_test.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'test_helper'
-require 'csv'
 module Api
   module V1
     class BooksControllerTest < ActionController::TestCase
@@ -93,35 +92,6 @@ module Api
           assert_equal body['name'].size, book.name.size
           # assert_equal body['query_doc_pairs'][0]['query'], book.query_doc_pairs[0].query_text
           # assert_nil body['query_doc_pairs'][0]['document_fields']
-        end
-      end
-
-      describe 'Exporting a book in basic csv format' do
-        let(:book)        { books(:james_bond_movies) }
-        let(:judgement)   { judgements(:jbm_qdp10_judgement) }
-        let(:doug)        { users(:doug) }
-        let(:random_user) { users(:random) }
-
-        test 'returns book w/ query doc pairs and judgement info' do
-          get :show, params: { id: book.id, format: :csv }
-
-          assert_response :ok
-          csv = CSV.parse(response.body, headers: true)
-          assert_equal 'Best Bond Ever', csv[0]['query']
-          assert_equal 'GeorgeLazenby', csv[0]['docid']
-          assert_equal '3.0', csv[3]['Doug Turnbull']
-
-          assert_not_includes csv.headers, 'Unknown'
-        end
-
-        test 'handles a rating that is not associated with a user, and adds Unknown' do
-          judgement.user = nil
-          judgement.save!
-          get :show, params: { id: book.id, format: :csv }
-
-          assert_response :ok
-          csv = CSV.parse(response.body, headers: true)
-          assert_includes csv.headers, 'anonymous'
         end
       end
     end
