+++ test/controllers/api/v1/judgements_controller_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'test_helper'
-
+require 'csv'
 module Api
   module V1
     class JudgementsControllerTest < ActionController::TestCase
@@ -34,6 +34,35 @@ module Api
             }
             assert_response :ok
           end
+        end
+      end
+
+      describe 'Listing judgements for a book in basic csv format' do
+        let(:book)        { books(:james_bond_movies) }
+        let(:judgement)   { judgements(:jbm_qdp10_judgement) }
+        let(:doug)        { users(:doug) }
+        let(:random_user) { users(:random) }
+
+        test 'returns book w/ query doc pairs and judgement info' do
+          get :index, params: { book_id: book.id, format: :csv }
+
+          assert_response :ok
+          csv = CSV.parse(response.body, headers: true)
+          assert_equal 'Best Bond Ever', csv[0]['query']
+          assert_equal 'GeorgeLazenby', csv[0]['docid']
+          assert_equal '3.0', csv[3]['Doug Turnbull']
+
+          assert_not_includes csv.headers, 'Unknown'
+        end
+
+        test 'handles a rating that is not associated with a user, and adds Unknown' do
+          judgement.user = nil
+          judgement.save!
+          get :index, params: { book_id: book.id, format: :csv }
+
+          assert_response :ok
+          csv = CSV.parse(response.body, headers: true)
+          assert_includes csv.headers, 'Anonymous'
         end
       end
     end
